+++ crates/common_enums/src/connector_enums.rs
@@ -132,6 +132,7 @@ pub enum RoutableConnectors {
     Recurly,
     Redsys,
     Riskified,
+    // Santander,
     Shift4,
     Signifyd,
     Square,
@@ -292,6 +293,7 @@ pub enum Connector {
     Razorpay,
     Recurly,
     Redsys,
+    // Santander,
     Shift4,
     Square,
     Stax,
@@ -463,6 +465,7 @@ impl Connector {
             | Self::Rapyd
             | Self::Recurly
             | Self::Redsys
+            // | Self::Santander
             | Self::Shift4
             | Self::Square
             | Self::Stax
@@ -624,6 +627,7 @@ impl From<RoutableConnectors> for Connector {
             RoutableConnectors::Recurly => Self::Recurly,
             RoutableConnectors::Redsys => Self::Redsys,
             RoutableConnectors::Riskified => Self::Riskified,
+            // RoutableConnectors::Santander => Self::Santander,
             RoutableConnectors::Shift4 => Self::Shift4,
             RoutableConnectors::Signifyd => Self::Signifyd,
             RoutableConnectors::Square => Self::Square,
@@ -740,6 +744,7 @@ impl TryFrom<Connector> for RoutableConnectors {
             Connector::Rapyd => Ok(Self::Rapyd),
             Connector::Razorpay => Ok(Self::Razorpay),
             Connector::Riskified => Ok(Self::Riskified),
+            // Connector::Santander => Ok(Self::Santander),
             Connector::Shift4 => Ok(Self::Shift4),
             Connector::Signifyd => Ok(Self::Signifyd),
             Connector::Square => Ok(Self::Square),