+++ crates/hyperswitch_connectors/src/default_implementations.rs
@@ -1,9 +1,29 @@
+#[cfg(feature = "dummy_connector")]
+use common_enums::{CallConnectorAction, PaymentAction};
 // impl api::PaymentIncrementalAuthorization for Helcim {}
 // impl api::ConnectorCustomer for Helcim {}
 // impl api::PaymentsPreProcessing for Helcim {}
 // impl api::PaymentReject for Helcim {}
 // impl api::PaymentApprove for Helcim {}
 use common_utils::errors::CustomResult;
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+use hyperswitch_domain_models::router_flow_types::{
+    BillingConnectorInvoiceSync, BillingConnectorPaymentsSync, RecoveryRecordBack,
+};
+#[cfg(feature = "dummy_connector")]
+use hyperswitch_domain_models::router_request_types::authentication::{
+    ConnectorAuthenticationRequestData, ConnectorPostAuthenticationRequestData, PreAuthNRequestData,
+};
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+use hyperswitch_domain_models::router_request_types::revenue_recovery::{
+    BillingConnectorInvoiceSyncRequest, BillingConnectorPaymentsSyncRequest,
+    RevenueRecoveryRecordBackRequest,
+};
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+use hyperswitch_domain_models::router_response_types::revenue_recovery::{
+    BillingConnectorInvoiceSyncResponse, BillingConnectorPaymentsSyncResponse,
+    RevenueRecoveryRecordBackResponse,
+};
 #[cfg(feature = "frm")]
 use hyperswitch_domain_models::{
     router_flow_types::fraud_check::{Checkout, Fulfillment, RecordReturn, Sale, Transaction},
@@ -21,12 +41,6 @@ use hyperswitch_domain_models::{
     },
     router_request_types::PayoutsData,
     router_response_types::PayoutsResponseData,
-};
-#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
-use hyperswitch_domain_models::{
-    router_flow_types::revenue_recovery as recovery_router_flows,
-    router_request_types::revenue_recovery as recovery_request,
-    router_response_types::revenue_recovery as recovery_response,
 };
 use hyperswitch_domain_models::{
     router_flow_types::{
@@ -79,6 +93,12 @@ use hyperswitch_interfaces::api::payouts::{
 };
 #[cfg(all(feature = "v2", feature = "revenue_recovery"))]
 use hyperswitch_interfaces::api::revenue_recovery as recovery_traits;
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+use hyperswitch_interfaces::api::revenue_recovery::{
+    BillingConnectorInvoiceSyncIntegration, BillingConnectorPaymentsSyncIntegration,
+};
+#[cfg(feature = "dummy_connector")]
+use hyperswitch_interfaces::api::ConnectorVerifyWebhookSource;
 use hyperswitch_interfaces::{
     api::{
         self,
@@ -201,6 +221,7 @@ default_imp_for_authorize_session_token!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -323,6 +344,7 @@ default_imp_for_calculate_tax!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -412,6 +434,7 @@ default_imp_for_session_update!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stripe,
@@ -534,6 +557,7 @@ default_imp_for_post_session_tokens!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Square,
@@ -654,6 +678,7 @@ default_imp_for_create_order!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Square,
@@ -777,6 +802,7 @@ default_imp_for_update_metadata!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Square,
@@ -916,6 +942,7 @@ default_imp_for_complete_authorize!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Stax,
     connectors::Square,
@@ -1036,6 +1063,7 @@ default_imp_for_incremental_authorization!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1156,6 +1184,7 @@ default_imp_for_create_customer!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Square,
@@ -1263,6 +1292,7 @@ default_imp_for_connector_redirect_response!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1372,6 +1402,7 @@ default_imp_for_pre_processing_steps!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Stax,
     connectors::Square,
@@ -1492,6 +1523,7 @@ default_imp_for_post_processing_steps!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1616,6 +1648,7 @@ default_imp_for_approve!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1740,6 +1773,7 @@ default_imp_for_reject!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1863,6 +1897,7 @@ default_imp_for_webhook_source_verification!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -1986,6 +2021,7 @@ default_imp_for_accept_dispute!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2108,6 +2144,7 @@ default_imp_for_submit_evidence!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2229,6 +2266,7 @@ default_imp_for_defend_dispute!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2360,6 +2398,7 @@ default_imp_for_file_upload!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2469,6 +2508,7 @@ default_imp_for_payouts!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Square,
@@ -2588,6 +2628,7 @@ default_imp_for_payouts_create!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2710,6 +2751,7 @@ default_imp_for_payouts_retrieve!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2833,6 +2875,7 @@ default_imp_for_payouts_eligibility!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -2951,6 +2994,7 @@ default_imp_for_payouts_fulfill!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -3072,6 +3116,7 @@ default_imp_for_payouts_cancel!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -3194,6 +3239,7 @@ default_imp_for_payouts_quote!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -3317,6 +3363,7 @@ default_imp_for_payouts_recipient!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -3440,6 +3487,7 @@ default_imp_for_payouts_recipient_account!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -3564,6 +3612,7 @@ default_imp_for_frm_sale!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stax,
     connectors::Square,
@@ -3688,6 +3737,7 @@ default_imp_for_frm_checkout!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stax,
     connectors::Square,
@@ -3812,6 +3862,7 @@ default_imp_for_frm_transaction!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stax,
     connectors::Square,
@@ -3936,6 +3987,7 @@ default_imp_for_frm_fulfillment!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stax,
     connectors::Square,
@@ -4060,6 +4112,7 @@ default_imp_for_frm_record_return!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stripe,
     connectors::Stax,
@@ -4179,6 +4232,7 @@ default_imp_for_revoking_mandates!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -4301,6 +4355,7 @@ default_imp_for_uas_pre_authentication!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stripe,
@@ -4422,6 +4477,7 @@ default_imp_for_uas_post_authentication!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stripe,
@@ -4544,6 +4600,7 @@ default_imp_for_uas_authentication_confirmation!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -4657,6 +4714,7 @@ default_imp_for_connector_request_id!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -4774,6 +4832,7 @@ default_imp_for_fraud_check!(
     connectors::Razorpay,
     connectors::Recurly,
     connectors::Redsys,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Stax,
     connectors::Square,
@@ -4919,6 +4978,7 @@ default_imp_for_connector_authentication!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -5039,6 +5099,7 @@ default_imp_for_uas_authentication!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -5154,6 +5215,7 @@ default_imp_for_revenue_recovery! {
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stax,
@@ -5186,9 +5248,9 @@ macro_rules! default_imp_for_billing_connector_payment_sync {
         $(  impl recovery_traits::BillingConnectorPaymentsSyncIntegration for $path::$connector {}
             impl
                 ConnectorIntegration<
-                recovery_router_flows::BillingConnectorPaymentsSync,
-                recovery_request::BillingConnectorPaymentsSyncRequest,
-                recovery_response::BillingConnectorPaymentsSyncResponse
+                BillingConnectorPaymentsSync,
+                BillingConnectorPaymentsSyncRequest,
+                BillingConnectorPaymentsSyncResponse
             > for $path::$connector
             {}
         )*
@@ -5279,6 +5341,7 @@ default_imp_for_billing_connector_payment_sync!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stripe,
@@ -5310,9 +5373,9 @@ macro_rules! default_imp_for_revenue_recovery_record_back {
         $( impl recovery_traits::RevenueRecoveryRecordBack for $path::$connector {}
             impl
             ConnectorIntegration<
-            recovery_router_flows::RecoveryRecordBack,
-            recovery_request::RevenueRecoveryRecordBackRequest,
-            recovery_response::RevenueRecoveryRecordBackResponse
+            RecoveryRecordBack,
+            RevenueRecoveryRecordBackRequest,
+            RevenueRecoveryRecordBackResponse
             > for $path::$connector
             {}
         )*
@@ -5402,6 +5465,7 @@ default_imp_for_revenue_recovery_record_back!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Shift4,
     connectors::Signifyd,
     connectors::Stripe,
@@ -5433,9 +5497,9 @@ macro_rules! default_imp_for_billing_connector_invoice_sync {
         $( impl recovery_traits::BillingConnectorInvoiceSyncIntegration for $path::$connector {}
             impl
             ConnectorIntegration<
-            recovery_router_flows::BillingConnectorInvoiceSync,
-            recovery_request::BillingConnectorInvoiceSyncRequest,
-            recovery_response::BillingConnectorInvoiceSyncResponse
+            BillingConnectorInvoiceSync,
+            BillingConnectorInvoiceSyncRequest,
+            BillingConnectorInvoiceSyncResponse
             > for $path::$connector
             {}
         )*
@@ -5525,6 +5589,7 @@ default_imp_for_billing_connector_invoice_sync!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -5642,6 +5707,7 @@ default_imp_for_external_vault!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -5765,6 +5831,7 @@ default_imp_for_external_vault_insert!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -5888,6 +5955,7 @@ default_imp_for_external_vault_retrieve!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -6011,6 +6079,7 @@ default_imp_for_external_vault_delete!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -6133,6 +6202,7 @@ default_imp_for_external_vault_create!(
     connectors::Recurly,
     connectors::Redsys,
     connectors::Riskified,
+    connectors::Santander,
     connectors::Signifyd,
     connectors::Shift4,
     connectors::Stax,
@@ -6158,3 +6228,506 @@ default_imp_for_external_vault_create!(
     connectors::Zen,
     connectors::Zsl
 );
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentsCompleteAuthorize for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<CompleteAuthorize, CompleteAuthorizeData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorVerifyWebhookSource for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        VerifyWebhookSource,
+        VerifyWebhookSourceRequestData,
+        VerifyWebhookSourceResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorCustomer for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<CreateConnectorCustomer, ConnectorCustomerData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorRedirectResponse for connectors::DummyConnector<T> {
+    fn get_flow_type(
+        &self,
+        _query_params: &str,
+        _json_payload: Option<serde_json::Value>,
+        _action: PaymentAction,
+    ) -> CustomResult<CallConnectorAction, ConnectorError> {
+        Ok(CallConnectorAction::Trigger)
+    }
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorTransactionId for connectors::DummyConnector<T> {}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> Dispute for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> AcceptDispute for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Accept, AcceptDisputeRequestData, AcceptDisputeResponse>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> FileUpload for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UploadFile for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Upload, UploadFileRequestData, UploadFileResponse>
+    for connectors::DummyConnector<T>
+{
+}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> RetrieveFile for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Retrieve, RetrieveFileRequestData, RetrieveFileResponse>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> SubmitEvidence for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Evidence, SubmitEvidenceRequestData, SubmitEvidenceResponse>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> DefendDispute for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Defend, DefendDisputeRequestData, DefendDisputeResponse>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentsPreProcessing for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<PreProcessing, PaymentsPreProcessingData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentsPostProcessing for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<PostProcessing, PaymentsPostProcessingData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> api::Payouts for connectors::DummyConnector<T> {}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutCreate for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoCreate, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutSync for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoSync, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutEligibility for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoEligibility, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutFulfill for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoFulfill, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutCancel for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoCancel, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutQuote for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoQuote, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutRecipient for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoRecipient, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PayoutRecipientAccount for connectors::DummyConnector<T> {}
+#[cfg(feature = "payouts")]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<PoRecipientAccount, PayoutsData, PayoutsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentApprove for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Approve, PaymentsApproveData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentReject for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<Reject, PaymentsRejectData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheck for connectors::DummyConnector<T> {}
+
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheckSale for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> ConnectorIntegration<Sale, FraudCheckSaleData, FraudCheckResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheckCheckout for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> ConnectorIntegration<Checkout, FraudCheckCheckoutData, FraudCheckResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheckTransaction for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8>
+    ConnectorIntegration<Transaction, FraudCheckTransactionData, FraudCheckResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheckFulfillment for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8>
+    ConnectorIntegration<Fulfillment, FraudCheckFulfillmentData, FraudCheckResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8> FraudCheckRecordReturn for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "frm", feature = "dummy_connector"))]
+impl<const T: u8>
+    ConnectorIntegration<RecordReturn, FraudCheckRecordReturnData, FraudCheckResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentIncrementalAuthorization for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        IncrementalAuthorization,
+        PaymentsIncrementalAuthorizationData,
+        PaymentsResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorMandateRevoke for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<MandateRevoke, MandateRevokeRequestData, MandateRevokeResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorPreAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorPreAuthenticationVersionCall for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorPostAuthentication for connectors::DummyConnector<T> {}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        Authentication,
+        ConnectorAuthenticationRequestData,
+        AuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<PreAuthentication, PreAuthNRequestData, AuthenticationResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        PreAuthenticationVersionCall,
+        PreAuthNRequestData,
+        AuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        PostAuthentication,
+        ConnectorPostAuthenticationRequestData,
+        AuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentAuthorizeSessionToken for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<AuthorizeSessionToken, AuthorizeSessionTokenData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> TaxCalculation for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<CalculateTax, PaymentsTaxCalculationData, TaxCalculationResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentSessionUpdate for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<SdkSessionUpdate, SdkPaymentsSessionUpdateData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentPostSessionTokens for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<PostSessionTokens, PaymentsPostSessionTokensData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentsCreateOrder for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<CreateOrder, CreateOrderRequestData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> PaymentUpdateMetadata for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<UpdateMetadata, PaymentsUpdateMetadataData, PaymentsResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UasPreAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UnifiedAuthenticationService for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        PreAuthenticate,
+        UasPreAuthenticationRequestData,
+        UasAuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UasPostAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        PostAuthenticate,
+        UasPostAuthenticationRequestData,
+        UasAuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UasAuthenticationConfirmation for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        AuthenticationConfirmation,
+        UasConfirmationRequestData,
+        UasAuthenticationResponseData,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> UasAuthentication for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<Authenticate, UasAuthenticationRequestData, UasAuthenticationResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> RevenueRecovery for connectors::DummyConnector<T> {}
+
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> api::revenue_recovery::BillingConnectorPaymentsSyncIntegration
+    for connectors::DummyConnector<T>
+{
+}
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        BillingConnectorPaymentsSync,
+        BillingConnectorPaymentsSyncRequest,
+        BillingConnectorPaymentsSyncResponse,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> api::revenue_recovery::RevenueRecoveryRecordBack
+    for connectors::DummyConnector<T>
+{
+}
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        RecoveryRecordBack,
+        RevenueRecoveryRecordBackRequest,
+        RevenueRecoveryRecordBackResponse,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> BillingConnectorInvoiceSyncIntegration for connectors::DummyConnector<T> {}
+#[cfg(all(feature = "v2", feature = "revenue_recovery"))]
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<
+        BillingConnectorInvoiceSync,
+        BillingConnectorInvoiceSyncRequest,
+        BillingConnectorInvoiceSyncResponse,
+    > for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalVault for connectors::DummyConnector<T> {}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalVaultInsert for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<ExternalVaultInsertFlow, VaultRequestData, VaultResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalVaultRetrieve for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8>
+    ConnectorIntegration<ExternalVaultRetrieveFlow, VaultRequestData, VaultResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalVaultDelete for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<ExternalVaultDeleteFlow, VaultRequestData, VaultResponseData>
+    for connectors::DummyConnector<T>
+{
+}
+
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ExternalVaultCreate for connectors::DummyConnector<T> {}
+#[cfg(feature = "dummy_connector")]
+impl<const T: u8> ConnectorIntegration<ExternalVaultCreateFlow, VaultRequestData, VaultResponseData>
+    for connectors::DummyConnector<T>
+{
+}