+++ crates/hyperswitch_domain_models/src/configs.rs
@@ -96,6 +96,7 @@ pub struct Connectors {
     pub recurly: ConnectorParams,
     pub redsys: ConnectorParams,
     pub riskified: ConnectorParams,
+    pub santander: ConnectorParams,
     pub shift4: ConnectorParams,
     pub signifyd: ConnectorParams,
     pub square: ConnectorParams,